ptfs: check mount id by using statx()
Import implementation of `statx()` and validate mount id when getting
mount fd.
